Don't use the puppetlabs/nginx module
The module is a good choice for simple configurations, but ours is
quite complicated, and the module doesn't even support rewriting.

This nginx module seems to be more powerful:
https://github.com/BenoitCattie/puppet-nginx

But it makes sense to manually install nginx and place the
configuration files for now, at least until the setup works. It's
easier to port that way.