Install Nginx from the official repository
I had to remove the apt update from the pre stage and add a dependency
chain with Package instead, which is the cleaner solution anyway. With
apt update running in the pre stage, we couldn't add additional
sources using the apt package.

Also changed the server configuration of Discourse to make sure we
listen on port 80 for ipv4, Nginx 1.4.1 requires that.