Noissue - Support puppet agent --tags
Some times we wish to constrain the puppet catalog run to only update certain tags.
Although this is mostly considered bad practice, it often becomes necessary with
automatic puppet agents disabled.

Note that we add the -t to get prompted for sudo when NOPASSWD has not been configured:
Currently only "clean" puppet agent runs with --noop and/or --test have NOPASSWD.

It would probably be fine to allow all puppet agent arguments in modules/adblockplus/files/sudoers/puppet
but we can do that in a later change, if deemed safe.

Also add a print statement for showing what commmand will be run remotely.