#3299 - Add URI redirects for missing translations
This change is related to the previous commit for #3421.

After this change, If someone has manually chosen Mexican (es-MX), and they
surf to a page that does not have a Mexican, but a Spanish (es) version
available, they will be flipped to Spanish and not return when later surfing to
a page available in Mexican.

The existing behavior was to just show 404s in these cases, which would be
terrible experience with all those half-complete translation sets we have
currently.

We could check for an available translation on every request, and flip them back
to Mexican when possible, but that will be counter-intuitive to those who have
manually chosen a different language in the top-right selector on the page.

The optimal solution would be to keep their last chosen language stored in
session and say "You're currently surfing the Mexican site but this page only
has Spanish translation available", but that would mean revisiting privacy
policy and server mechanics, which is out of scope for this commit.