#5020 - Bridge trivial legacy redirects to modern redirect service
Note that this change requires that it is first deployed to eyeo.to before being deployed to adblockplus.org. We could also consider splitting this into two separate, safer changes.

Also note: The changes have not undergone rigurous testing yet! (not so easy to test tho, which could be another argument for splitting it into two separate changes).

Finally note that this commit applies cleanly only after https://codereview.adblockplus.org/29613809/ landing first, but it can easily be reworked to apply without it.