HD-1398 - Fix bug affecting $lang parameter for share.adblockplus.org redirect
Bug fix for modules/web/templates/adblockplus.org.conf.erb where the $lang parameter gets reset to a default value of "en" if a local file for a localized page cannot be found.
However, this also affects external redirects like the one to share.adblockplus.org, so this redirect needs to happen before that fallback to "en" for non-existing local pages kicks in.